Fix a couple of nits picked up by gcc-4.5, which is more warningful than 4.4.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@10876

+++ coregrind/launcher-linux.c
@@ -135,7 +135,8 @@ static const char *select_platform(const char *clientname)
       return NULL;
    }
 
-   VG_(debugLog)(2, "launcher", "read %ld bytes from '%s'\n", n_bytes, clientname);
+   VG_(debugLog)(2, "launcher", "read %ld bytes from '%s'\n",
+                    (long int)n_bytes, clientname);
 
    if (header[0] == '#' && header[1] == '!') {
       int i = 2;
